pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = "poetry project"
authors = ["Amit Sahoo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.3"
python-dotenv = "^1.0.0"
fastapi = {extras = ["all"], version = "^0.95.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
